refactor: move status bar creation from extension.ts to command layer
